+++ src/Controller/Api/v1/Percentage/PercentageController.php
@@ -7,15 +7,18 @@ use App\Controller\Api\v1\Percentage\Input\CreateData;
 use App\Controller\Api\v1\Percentage\Input\ReadData;
 use App\Controller\Api\v1\Percentage\Input\UpdateData;
 use App\Entity\Percentage;
+use App\Exception\AccessDeniedException;
 use App\Exception\BadRequestException;
 use App\Exception\EntityNotFoundException;
 use App\Manager\PercentageManager;
-use App\Repository\PercentageRepository;
+use App\Manager\StaffManager;
+use App\Service\PercentageService;
 use Exception;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpKernel\Attribute\MapQueryString;
 use Symfony\Component\HttpKernel\Attribute\MapRequestPayload;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\User\UserInterface;
 
 #[Route(path: '/api/v1/percentage')]
 class PercentageController extends AbstractController
@@ -23,7 +26,7 @@ class PercentageController extends AbstractController
 
     public function __construct(
         private readonly PercentageManager $percentageManager,
-        private readonly PercentageRepository $percentageRepository
+        private readonly PercentageService $percentageService
     ) {
 
     }
@@ -34,6 +37,7 @@ class PercentageController extends AbstractController
     #[Route(path: '', methods: ['POST'])]
     public function create(#[MapRequestPayload] CreateData $dto): Percentage
     {
+        $this->denyAccessUnlessGranted(StaffManager::ROLE_STAFF);
         return $this->percentageManager->create($dto);
     }
 
@@ -43,6 +47,7 @@ class PercentageController extends AbstractController
     #[Route(path: '', methods: ['PATCH'])]
     public function update(#[MapRequestPayload] UpdateData $dto): Percentage
     {
+        $this->denyAccessUnlessGranted(StaffManager::ROLE_STAFF);
         return $this->percentageManager->update($dto);
     }
 
@@ -52,16 +57,18 @@ class PercentageController extends AbstractController
     #[Route(path: '/{id}', requirements: ['id' => '\d+'], methods: ['DELETE'])]
     public function delete(int $id)
     {
+        $this->denyAccessUnlessGranted(StaffManager::ROLE_ADMIN);
         return $this->percentageManager->delete($id);
     }
 
     /**
      * @throws EntityNotFoundException
      * @throws BadRequestException
+     * @throws AccessDeniedException
      */
     #[Route(path: '', methods: ['GET'])]
-    public function read(#[MapQueryString] ReadData $dto): EntityCollection
+    public function read(#[MapQueryString] ReadData $dto, UserInterface $user): EntityCollection
     {
-        return $this->percentageRepository->getPercentages($dto);
+        return $this->percentageService->getPercentages($dto, $user);
     }
 }
